add fslgit recipe for QEMU 1.4
Note: in this recipe --cross-prefix is added so that QEMU
configure will properly detect the endianness of the
target (host).

The use of cross-prefix then adds the need to explicly
set PKG_CONFIG so that configure gets the native pkg-config,
so we add a do_configure_prepend() to set this.

Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>
Change-Id: I1786f437d626c689f3e6d8a32e46dfa47fe2528e
Reviewed-on: http://git.am.freescale.net:8181/1189
Reviewed-by: Luo Zhenhua-B19537 <zhenhua.luo@freescale.com>
Tested-by: Luo Zhenhua-B19537 <zhenhua.luo@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/2095