Disable Busybox SUID splitting
Commit [1] introduces the ability to split the Busybox binary
in 2 individual binaries, one containing applets that require
SUID and the other containing ones that don't. This solves
the problem of not having a full SUID-ed busybox for just a
few tools, such as ping, traceroute, login, etc. The option
has been introduced by default in poky/meta/recipes-core.

This impacts scenarios that make use of Busybox indirectly,
e.g. LXC. The Busybox container initialization script looks for
a single busybox binary on the host, and requires certain
applets to be included in it - login, passwd. The current
status is that the script finds the busybox.nosuid binary which
doesn't include these applets, and fails.

Modifying the LXC script to look for busybox.suid as well is
a specific change pertaining only to the LXC built by Yocto,
thus it won't be accepted upstream. Making the change to the
meta-virtualization layer would mean maintaining it as LXC
versions increment.

Pushing this fix here, since it's in a portion with QorIQ
specific settings. Please apply if there's no specific need
elsewhere for Busybox to be split into suid and non-suid.

Additional argument: the size of busybox is smaller when
compiled in a single binary - at least when it's built
statically, which is a requirement when used with LXC.

[1] http://git.yoctoproject.org/cgit/cgit.cgi/poky/commit/meta?id=680ff5552d6953c30ce1a8ff127d4aea71e81a17

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@freescale.com>
Change-Id: Id3fce87644e6ba415b20eb777c5408fd119f6ba7
Reviewed-on: http://git.am.freescale.net:8181/10735
Reviewed-by: Zhenhua Luo <zhenhua.luo@freescale.com>
Tested-by: Zhenhua Luo <zhenhua.luo@freescale.com>