openssl: add missing patches for yocto recipe
Fixing the openssl snapshot to the digest of Freescale OpenSSL
repository has the side-effect of removing several patches that come
with yocto which are necessary for a correct package build.
Since these patches are part of yocto, we add them back.

This fixes a build failure of openvswitch which depends on OpenSSL.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>
Change-Id: I99f6006aed805bffa12439feebe25d8600ac6352
Reviewed-on: http://git.am.freescale.net:8181/34633
Reviewed-by: Zhenhua Luo <zhenhua.luo@freescale.com>
Tested-by: Zhenhua Luo <zhenhua.luo@freescale.com>