fsl-toolchain: Remove broken meta-environment
Currently meta-environment packages the toolchain environment files and
they get included in the SDK but are broken. At the meantime, the same
fuction was duplicated in populate_sdk_base.

Upstream already fixed this now. Comparing to backport several upstream
commits, it is safer to remove the broken meta-environment from install
list.

Signed-off-by: Ting Liu <ting.liu@freescale.com>
Change-Id: Ica9987b0c39268390e0513e087de09c39e3e09d3
Reviewed-on: http://git.am.freescale.net:8181/35337
Reviewed-by: Zhenhua Luo <zhenhua.luo@freescale.com>
Tested-by: Zhenhua Luo <zhenhua.luo@freescale.com>